requirements_dev.txt
pip==19.2.3
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
pytest==6.2.4
black==21.7b0
requests-mock==1.10.0
pytest-dotenv==0.5.2
pytest-mock==3.10.0
